+++ CHANGELOG.md
@@ -6,8 +6,9 @@ Converters:
 * Removed the setRegistry function
 * getReturn now returns the fee as a separate return value
 * Converter owner can no longer withdraw connector tokens while the converter is active
-* Converter owner can no longer transfer the token ownership
+* Converter owner can no longer transfer the token ownership once the converter is active
 * Added a dedicated `upgrade` function for easier upgrades
+* Added a `fund` function for increasing liquidity atomically
 * Added a `liquidate` function for decreasing liquidity atomically even when conversions are disabled
 
 Registry: